Refactor hero page logic to enhance maintainability
Streamlined hero retrieval and navigation by moving logic to _getPage.ts, thereby simplifying the HeroPage component. This ensures better separation of concerns and easier future modifications.
