+++ frontend-next-migration/src/app/[lng]/(helper)/heroes/[slug]/page.tsx
@@ -1,10 +1,6 @@
 import { HeroPage as PreparedHeroPage } from '@/preparedPages/HeroesPages';
-import { useServerTranslation } from '@/shared/i18n';
-import { heroes, HeroManager, HeroSlug, HeroWithGroup } from '@/entities/Hero';
-import { RoutePaths } from '@/shared/appLinks/RoutePaths';
 import { notFound } from 'next/navigation';
-// import {withPageData ,createMetadataGenerator} from "@/app/_helpers";
-// import {_getPage} from "./_getPage";
+import { _getPage } from './_getPage';
 
 interface Props extends DefaultAppRouterProps {
     params: DefaultAppRouterProps['params'] & {
@@ -12,33 +8,24 @@ interface Props extends DefaultAppRouterProps {
     };
 }
 
+//todo add seo, try with it:
+// import {withPageData ,createMetadataGenerator} from "@/app/_helpers";
+// import {_getPage} from "./_getPage";
+
 export default async function HeroPage({ params }: Props) {
     const { slug, lng } = params;
-    const { t } = await useServerTranslation(lng, 'heroes');
-    const heroManager = new HeroManager(t);
-    const currentHero = heroManager.getHeroBySlug(slug as HeroSlug);
-    if (!currentHero) {
+    const pageProps = await _getPage(lng, slug);
+
+    if (!pageProps) {
         return notFound();
     }
-    const heroes = heroManager.getAllHeroes();
-    const prevHero =
-        heroManager.getHeroBeforeSpecificHero(currentHero.id) || (heroes.at(-1) as HeroWithGroup);
-    const nextHero =
-        heroManager.getHeroAfterSpecificHero(currentHero.id) || (heroes.at(0) as HeroWithGroup);
-    const prevHeroLink = generateHeroLink(prevHero.slug);
-    const nextHeroLink = generateHeroLink(nextHero.slug);
-
     return (
         <div>
             <PreparedHeroPage
-                newSelectedHero={currentHero}
-                prevHeroLink={prevHeroLink}
-                nextHeroLink={nextHeroLink}
+                newSelectedHero={pageProps.page.selectedHero}
+                prevHeroLink={pageProps.page.prevHeroLink}
+                nextHeroLink={pageProps.page.nextHeroLink}
             />
         </div>
     );
-}
-
-function generateHeroLink(heroSlug: string): string {
-    return RoutePaths.HEROES_ONE.replace(':slug', heroSlug);
 }
